backend/requirements.txt
flask>=3.0.0,<4
google-genai>=1.0.0,<2
openai>=1.40.0,<2
python-dotenv>=1.0.0,<2
requests>=2.31.0,<3
pydantic>=2.0.0,<3
boto3>=1.34.0,<2
